Adjust dependencies of the anaconda-webui package
Set minimum cockpit version to make sure we have cockpit-client
with correct WebKit API support.

Also require the correct WebKit API to be available as cockpit-client
does not have a hard requires on it itself due to its fallback mechanism,
which does not really work in the installation environment
(eq. we don't have a Firefox or Chrome browsers to fall back to).