Drop packit override for webui package build
The anaconda-webui sub-package is gated by the spec file use_cockpit macro.

So far we have changed this macro via packit to disable the
anaconda-webui package from being build in the official Fedora builds.

This is no longer necessary & we are ready for the anaconda-webui
package to be part of the official Anaconda builds, so lets drop
the override.