Do not pass rd.znet on to installed system unconditionally
Passing the rd.znet boot argument to the installed was implemented as a
workaround for bug https://bugzilla.redhat.com/show_bug.cgi?id=1087502,
commit 64fb106734413550985dcd568ab6b3baba76603e.

With this change, the rd.znet boot argument is no longer passed on to
the installed system unconditionally and the network device is
configured and activated after switchroot by udev/NetworkManager
(tested). When networking is needed in initramfs (like in case of
iSCSI), rd.znet is automatically added to the kernel command line of
the installed via a different mechanism (tested). Also the original
problem in bug 1087502 (installation via a VLAN network interface and
network availability on the installed system) was successfully proven
to be working fine with this change.

This change will also prevent confusing issues we have seen:
https://bugzilla.redhat.com/show_bug.cgi?id=1726060
https://bugzilla.redhat.com/show_bug.cgi?id=1966345