infra: Use the Bugzilla API key in the `makebumpver` script
The `~/.rhbzauth` configuration file is no longer supported. Use the
configuration file supported directly by `python-bugzilla` instead.
The support for username and password was removed. Use the API key instead.

The `makebumpver` script will ask for the API key interactively if none is
available. The key will be saved to a configuration file and reused next time.
Alternatively, you can create the `~/.bugzillarc` file with the API key
yourself with the following content:

    [bugzilla.redhat.com]
    api_key = <API-KEY>