infra: Use environment for releases
Github actions can not push to protected branches because it is not in the
list of permitted teams and members. (Currently, it can not be so by design.)

Therefore, use a token generated by someone who can push to protected
branches. Of course, the token must be generated by such a person in the first
place.

The idea here is that the token is gnerated to the bot account, InStallker.