infra: Push only the new tag and only if able to
Push the tag and commit both at once. For <reasons>, this can result in
successfully pushing the tag even if the commit won't go through.
The triggering condition is when someting was merged in the repo and the
checkout is no longer current by the time the push happens.
So make this push atomic.

We should also not push all tags using --tags. It's bad practice, because
it tries to push all local tags. Athough in this particular workflow we add
only the one new tag that we want, it is still a place that de facto
self-documents how to do this manually.