UI: Show error messages if the payload setup finishes early
When we register callbacks to signals of the payload manager in GUI and TUI,
it is possible that the payload manager is already finished with the payload
setup. In that case, we have to trigger the failed callback manually to set up
the error messages. Otherwise, the error messages won't be visible in the UI.