Sort RPM versions via rpm.labelCompare() and not via packaging.version.LegacyVersion()
Packaging 22+ removed the long-deprecated packaging.version.LegacyVersion class.
The packaging.version library is intended to parse Python (PEP 440) versions,
not arbitrary versions and definitively not RPM versions.

Even if LegacyVersion was still available, it may produce incorrect results
if the version contains ~, ^ or other characters with special meaning in RPM.

This assumes the Python rpm module is always available.
If that assumption is wrong
the logic from rpm.labelCompare() needs to be reimplemented instead.