Templatize pykickstart version in dracut code
The kickstart_version module needs to be put somewhere to import. Detecting
the python site-packages directory is apparently too variable, and the only
shortcut was in distutils which is to be removed. So let's just dump the
module next to the script and be done with it.

The module needs to be a separate file because of the templates.
The alternative is to make a template out of the whole script.

Also, add warning to templates about automatic RHEL kickstart version.