Move loop device pre-creation to container (#infra)
We have to pre-create loop devices because they are not available in the
containers. If they are created by `losetup` than the newly created loop
device is not visible inside the container. However, it still could be
created manually in the container and taken correctly by the losetup
tooling.

Thanks to the above discovery let's simplify our solution and create
loop devices inside the lorax-build script intead of manually before.