DeviceTreeViewer: Add path-id attribute to zfcp-attached SCSI disks
Currently, the UI prints the kernel device names for
multipath path members. Kernel device names are unpredictable and
non-persistent and thus hardly suitable as path identifier for users.

This is a preparation providing backend data to be used in a future commit
to show in the UI.

FCoE, iSCSI, and NVDIMM already have the attribute since commit
e74030ef4c62 ("Add the path-id attribute to the DBus structure for device
data"). Provide the same for zfcp-attached SCSI disks, which should always
be used with multipathing.

This change tolerates the absence of the id_path attribute if the
corresponding change adding the attribute in blivet does not exist.

Signed-off-by: Steffen Maier <maier@linux.ibm.com>