oemdrv: wait up to 5 seconds for disks to be handled
OEMDRV disk is used for both Driver Update and Kickstart.
When booting with the DVD and an additional OEMDRV disk, it appears that
the udev rule setting the script to fetch the kickstart may not run at
the time /dev/root was found on the DVD, causing the kickstart to not be
processed at all.
This patch delays 'settled' initqueue by 5 seconds by default when no
'inst.ks' is specified on the kernel command line, which is enough to
process the OEMDRV disk.
Because the same issue may appear with Driver Updates, that same delay
applies unless 'inst.dd' is explicitly specified.
Old behavior can be restored with adding 'inst.wait_for_disks=0' on the
kernel command line.

See RHBZ #2171811 for details and reproducer.

Signed-off-by: Renaud Métrich <rmetrich@redhat.com>