Fix translations of the pre-release warning dialog (#2165762)
Move the pre-release warning dialog to a new class. Thanks to that the UI
elements of the dialog will be created and translated when we need to show
the dialog. The locale is already set to the right value at that point, so
we don't have to retranslate the text manually.

Related: rhbz#2127473

(cherry picked from commit 64219c599759bfec1e538fe30d4a6f6819a3c65d)