Fix translations of the pre-release warning dialog (#2165762)
Move the pre-release warning dialog to a new class. Thanks to that the UI
elements of the dialog will be created and translated when we need to show
the dialog. The locale is already set to the right value at that point, so
we don't have to retranslate the text manually.

Resolves: rhbz#2165762