fedora-welcome: Add back app icon
The skip-taskbar-hint is specific to X11. On wayland there is no
way of hiding the app from the top bar, overview or alt-tab, so
if we don't want a generic fallback icon, we must specify one in
the .desktop file.

This undos some of the changes from commit 20eca42c2e3e7f47f933.