fedora-welcome: Use libadwaita
libadwaita is GNOME's platform library on top of GTK, which
provides the platform styling, support for the dark style
preference and lots of convenience widgets for common design
patterns.

Simply switching to AdwApplication would be enough for styling
and dark-mode support, however we also leverage the StatusPage
widget to make the overall layout more standardized, and to
move the window title inside the window.