fedora-welcome: Default to light style
The primary use case for a dark style are content applications
like video- or image viewers/editors. This doesn't apply to
the welcome screen, so switch to the regular (light) variant.

As we now integrate with the desktop's dark-style support, users
who prefer an overall dark look can still enable it.