WebUI: do not force to manually remove updates.img
You update something in source code and you need to have a new
updates.img with new anaconda sources. However, the current solution is

make ../../updates.img

that means that if the file exists you have to manually remove it first.
Add a new make target which do that for you.