Add --strict mode to makeupdates script
Right now it's just for --add-rpm but could be extended.

Reason for this is that adding rpms could break and this check would
kill the process with meaningful message. The current situation is that
the process quitely ignore the issue and continues.

In my case I had 404 return code when downloading cockpit packages for
cockpit-ci local run. However, it was hard to spot that these were not
added so the resulting updates_image did not contained them. This is
pretty bad situation which is hard to find what happened.