Move ostreecontainer deps to install-img-deps
The dependencies should be in core package correctly but having it there
will drag in these packages to Live environment and also all systems
installed with Live ISO which still have Anaconda package.

Moving it to install-img-deps package will add the requirement only to
boot.iso which is the main use for this.

Related: rhbz#2125655