Fix wrong dracut timeout message
We added Anaconda error reporting when Dracut timeout during boot. If
the reason is known issue we will print that to a user. However, in case
it is not known the code failed with message:

... /run/anaconda/initrd_errors.txt: No such file or directory

To fix this just print "Reason unknown" and not execute rest of the
code.

(cherry picked from commit 87780c8ba530f10cc4e489c387fc34084f98afee)

Resolves: RHEL-5638