webui: use run-tests from test/common with some severe HACKS to make it work with anaconda Vm building script
test/run-tests is an extremely old and patched version of
test/common/run-tests, so that the global machine creation can use
anaconda's test/webui-testvm.py class intead of the bots/testvm.py
original.

This script was meant to be used as a temporary workaround but it lived
for more than it should have.
See https://issues.redhat.com/browse/INSTALLER-2866o

The problem is that this script is very outdated and it often blocks the
test/common version bumps.

This commit brings us one step closer in using the original test/common/run-tests directly.

It does that by overriding the bots/testvm.py with a symlink to
anaconda's custom script for VM creation.