exception: only attach existent and non-empty files (#2185827)
libreport barfs (in a non-fatal but scary way) if a file we try
to attach does not exist or is empty. Let's make sure all the
files exist and aren't empty before we try to attach them. In a
current F38+ live install, for instance, both packaging.log and
dnf.librepo.log are usually empty.

Signed-off-by: Adam Williamson <awilliam@redhat.com>