Increase the minimum EFI System Partition (ESP) size to 500MiB
Currently the default allocated space for the EFI System Partition (ESP) is
a minimum of 200 MiB and a maximum of 600MiB. This smaller size is more than
enough for storing the EFI binaries just needed to boot the machine.

The ESP is also used to deploy firmware updates which need 2x SPI flash size
(typically 64MB, soon 128MB) as free space. We're also soon going to modernize
how the bootloader is deployed, and this also is going to need more space again.

Microsoft mandates OEMs to allocate a minimum of 500MiB and so I think we should
just do the same. Also increase the maximum size to 2GiB to accommodate future
UKIs.

Signed-off-by: Richard Hughes <richard@hughsie.com>