Remove link to the release notes template.rst
Reason is that the CONTRIBUTING.rst file is used from two places.
Directly on GitHub and from our documentation. Unfortunately both of
these places are used from a different place

- GH has working directory on repository root
- documentation is builded from ./docs directory

As addition to this the current solution doesn't work from neither
because it's sphinx based which is not understanded by GH correctly.

To solve this just leave it on user to open the correct page.
Unfortunately I failed to found better solution.

Also fix outdated documentation which works with directory instead of
the release-notes.rst file.