GUI: Require 8 chars of LUKS passwords in FIPS mode
The underlying mechanism requires 8 bytes. To prevent encoding/decoding
problems, codepoint counting etc., we just get the length of the unicode
string as Python understands it. That is a crude approximation, but fails in
the right direction (enforces stronger passphrase), and appears to count
graphemes correctly for the scripts most likely to be used by well-meaning
human users.

For emoji, all guarantees are off, but there is no way to enter these. (Yet?)

Resolves: rhbz#2163497