Use inst.proxy in stage1 (#2177219)
Right now we parse `inst.proxy=` in stage2 but not using that directly in
the stage1 (Dracut). On the other hand, Dracut is using `proxy=` from the
kernel cmdline parameters and ignore `inst.proxy=`.

At the end we ended up with `inst.proxy=` is used when installing from
the repositories but not during boot to download stage2 image,
kickstart and similar. The `proxy=` was consumed by stage2 and stage1
but not recommended by the documentation to be used.

To keep the consistency with already existing `inst.noverifyssl` we will
add `proxy` to dracut when we found `inst.proxy` in stage1.

Related: rhbz#rhbz#2177219