Don't override blivet's preferred disk label type by default
https://github.com/rhinstaller/anaconda/pull/4232 made anaconda
prefer GPT disk labels by default. This was intended to implement
the Change called "Install Using GPT on x86_64 BIOS by Default":
https://fedoraproject.org/wiki/Changes/GPTforBIOSbyDefault
https://bugzilla.redhat.com/show_bug.cgi?id=2092091
However, it actually does more than that Change requested. It
prefers GPT in other cases - it's quite hard to parse out every
possible case this changes, but for example, it also prefers GPT
on ppc64le installs, which previously preferred MBR. It's also
just an odd approach, when blivet already goes to some lengths
specifically to provide a list of supported label types in
a preferred order. It makes much more sense for anaconda to keep
respecting the order of the list blivet provides, and adjust
blivet to list GPT first on x86_64. The config setting and
`inst.disklabel` boot option can still be used to force a
preference; this just makes the default config value unset rather
than 'gpt'.

This goes with this blivet PR that implements preferring GPT on
x86_64 (even on BIOS):
https://github.com/storaged-project/blivet/pull/1132

Signed-off-by: Adam Williamson <awilliam@redhat.com>