If nano is the default editor, use it for bug reports
This affects meh and libreport in text mode.

It is not possible to source the bash file, because we run python, not bash.
The file is hardcoded because there is no way to walk over the whole system
and find if nano is being used.

Doing it this way also preserves the ability to not have the file, and use
vi as before.

Resolves: rhbz#1889674