Import BlockDev from blivet instead of gi
We are planning to release a new major version of libblockdev soon
and because version needs to be specified when importing from gi
it would make the import unnecessarily complicated if we wanted to
keep it backwards compatible. Anaconda is already using blivet so
importing BlockDev from it instead of gi shouldn't change anything.