Revert "Download cockpit rpms during build"
This reverts commit 326cda0817ed2521716db6d1bc3dd90c165bfb23.

This commit that I revert here was hindering reproducability of the
updates.img.

We need to get the same version of anaconda-webui dependencies RPMs
every time, instead of these randomly/confusingly changing just because
a new update passed bodhi.

The build-rpms script will likely be obsolete soon, when anaconda will
start building their own ISO files, which will contains the
anaconda-webui deps. Till then let's have:

* The dependencies gated
* The downloaded rpms cached

The only downside of the change is that now we need to manually update
the versions of the deps, but it's much safer to gate this and the only
guarantee for a green CI.