Fix "proxy" boot option is printing inst. warning (#2177219)
We warn user that our proxy kernel command line option should be used
with the 'inst.' prefix. However, it is also processed by Dracut (and
maybe others) to set the curl calls correctly.

We are even generating proxy command for Dracut when we download stage2
based on the KS url with --proxy parameter.

For this command we should not warn user about the use of the 'inst.'
prefix because both variants are correct.

Related: rhbz#2177219