Solve permission error for labeler action on forks (#infra)
Fork has lower permissions as external contributor which in this case is
an issue because labeler needs to have also the `checks: write` set.
Unfortunately the pull_request can't get bigger permissions than the
owner of the action so to fix this switch to `pull_request_target`.

The `pull_request_target` is not recommended because it has max
permissions by default (we overriding this). Reason for the maximal
permissions are because the workflow is started not with token from fork
but with token from target repository. However, here we should be fine
because we don't even clone the repository nor comments.

Without this change PRs from forks are getting:
```
Error: HttpError: Resource not accessible by integration
Error: Resource not accessible by integration
```