infra: Add container to build Live ISO with web UI
This image will just build Fedora Workstation ISO with web UI in the
container. However, it won't use Anaconda source for the build that is
the next step.

This container is implemented by using livemedia-creator using qemu
which will get 80% of the host resources (vcpus and memory). The
--no-virt option is not usable because during installation of the
selinux-policy-targeted package by Anaconda the package starts to run
tools of selinux and that is recognized on host (with enabled selinux)
which will start doing pretty nasty things with the system. Because of
that I decided to go rather with the virt option of livemedia-creator
which spawns VM and run the installation in it.

The container contains latest Workstation kickstart from the
fedora-kickstarts project and store the resulting kickstart inside the
container which could allow us to gate it this way in the future.

The kickstart is modified during the run of the container to add the
anaconda-webui package.