webui: Disable minification
Disable minification when we build our JS code,
so that the code we ship in our RPMs can be debugged.

This should improve the debugging of runtime issues
on any media built from RPM packages as well as for
any cases where we apply updates images built from RPM packages.