Add live environment keyboard settings support
This configuration will be used if no keyboard layout is set by user and
Anaconda is running on supported live environment.

This solution is integrated to the already existing
`get_missing_keyboard_configuration` function before using the
DEFAULT_KEYBOARD layout (US).

Before this change the logic was:
- use configuration if set by user
- if only keyboard layout or virtual console is set convert the other by
localed
- if nothing set use DEFAULT_KEYBOARD

After this change:
- use configuration if set by user
- if on supported live environment, read the current keyboard
configuration
- if only keyboard layout or virtual console is set convert the other by
localed
- if nothing set use DEFAULT_KEYBOARD

Currently, only Gnome Shell is supported but we should extend this
solution ideally to all Wayland based spins. Wayland is hostile to
Anaconda because it don't allow us to control live keyboard
configuration. Live environment now have configuration of the keyboard
separate from the Anaconda configuration, which is used for the
installed system. That means that users can write passwords with other
layout than what is installed (especially problem for LUKS).