webui: unpack GetDiskFreeSpace and GetDiskTotalSpace from array
GetDiskFreeSpace and GetDiskTotalSpace returns an array with an actual
value at first index.
We then use this value to check if there is enough free space, e.g.:
"diskFreeSpace < requiredSize". While it may not have caused any issues
until now, since in JS comparison of array and value (e.g. [99] < 100
still evaluates as true), this is just invitation for bugs in the
future and needs to be fixed.