make: Add l10n config file with SHA variable, use
Note that getting only the desired SHA "downloaded" is complicated due to git
and GitHub choices. Cloning fetches and checks out both at once, which is what
the original code did. However, that is allowed only for named branches and
tags.

It is possible to configure the git *server* to allow this for arbitrary SHA,
but it is neither default nor recommended, so GitHub does not allow this.

So the actual series of calls is:
- clone the repo = fetch only one ref, actually does not matter which one
- fetch only the one commit that we want
- check out that commit