infra: Don't copy bots in container, we don't need them
This fixes the problem where having bots with a downloaded image failed the
container-rpms target because the image could not be copied for some reason.

The image did not have any reason to be copied in the container.

However, we remix the container files freely left and right, which is just
wonderful, so we need rsync in both containers.