webui: tests: extend the TestInstallationProgress test to include reboot
For making this possible add `--wait` option to virt-install invocation [0].
This now allows the VM to reboot when reboot is requested, whilst
previously it was just shutting down.

[0] https://listman.redhat.com/archives/virt-tools-list/2011-November/002869.html

Now that the VM reboots when reboot is requests it's not possible to
test the "Quit" button as on boot.iso the Quit button actually reboots
the VM into the unbootable disk (as the installation did not finish).
The later state will fail the test teardown, as the target VM is not
reachable.
Therefore the tests checking the 'Quit' button were removed.

Co-authored-by: Katerina Koukiou <kkoukiou@redhat.com>