infra: After image build, add comment for download
Reportedly, it is not possible to get an artifact link in the same *workflow*.
It is also not economical to trigger another workflow to bypass this
limitation, because the other workflow will have to determine anew which
artifact is the image.

Pointing to the run status page bypasses the need to link the actual file,
keeping the whole thing simple.