Don't directly copy webui files in makeupdates script
Originally we tried to copy webui files directly when creating an
updates image, like we do with our python, glade, C and other files.

But eventually it turned out to be easier to builds the RPMs and then
include those into the updates image instead.

But the original code was still in place, causing issues such as causing
an unexpected attempt to start the Web UI when trying to change
something in the Anaconda text interface.

So lets drop the direct copy of the webui files from the makeupdates
script, at least for now.

This should not influence the current Web UI workflow, which already
works with RPMs exclusively & should avoid issues when makeupdates is
used in non-Web-UI use cases.