Disable geolocation on the Fedora Workstation live image
We need to disable geolocation for the Fedora Workstation live image
or else this will clash with and possibly override the language
selection performed by Gnome Initial setup, which in this environment
runs before Anaconda is started.

As far as we can tell, the Fedora Workstation live image is the only
image using the Fedora Workstation configuration file, so it should
be safe to put it there.

Related: rhbz#2231339