webui: Launch storage utility via url-handler instead of cockpit.spawn
cockpit.spawn isn't a good command to run UI tools because it doesn't
know the user event that lead to it getting called. UI applications need
to be run with that sort of context in their environment or compositors
won't start them in the foreground. This is so if a user is typing, a
slow starting program doesn't steal focus inadvertently when it finally
starts.

Firefox has a mechanism for launching applications with the appropriate
context, however: Url Handlers.

This commit registers a url handler "anaconda-storage:" and changes
the Web UI to use it instead of cockpit.spawn.