Revert the ESP maximum size back to 600MiB
The original issue was that 200MiB was too small, and we only increased the
maximum for people experimenting with UKIs. Seeing as Anaconda chooses the
maximum size in more cases than we'd like just drop the maximum size back down
to 600MiB which is easily big enough for firmware updates.

This should fix several regressions like:

 * https://bugzilla.redhat.com/show_bug.cgi?id=2212121
 * https://bugzilla.redhat.com/show_bug.cgi?id=2214342

Signed-off-by: Richard Hughes <richard@hughsie.com>