webui: Hide close button in browser
When the user clicks the close button, it currently displays:

----
🌐 127.0.0.1

This page is asking you to confirm that you want to
leave — information you’ve entered may not be saved.

                      [ Stay on Page ] [ Leave Page ]
----

That is obviously not a great user experience.

Anaconda already has a way to abort the install and reboot, so
the close button isn't really adding much.

This commit hides the close button.