infra: Don't run infra check on merged PRs
Rebasing the head of a merged PR results in no changes (rebased == base),
which gives empty string for the difference, which lacks the infra prefix or
postfix.

The reason to actually check for this is that the "port to X" labels are
usually removed only after the PR has been merged. That triggers the action on
closed PRs, the only such case we have.