storage: Handle live images and inst.sdboot
Live images are already installed with grub2 and cannot be changed on
the fly to systemd-boot (well at least without a lot of extra
work). In case the user passed one of the systemd-boot options to
anaconda, we should be a bit more graceful than failing to run the
updateloader entries script as currently happens.

This change, when run on non webUI installs results in a popup
notifying the user there is an error in the bootloader configuration
and asking if they would like to ignore it.

Resolves: rhbz#2234638
Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>