Simplify keyboard layout handling, rely on localed more
There are a couple of specific bugs in the current code:

https://bugzilla.redhat.com/show_bug.cgi?id=2238854
https://bugzilla.redhat.com/show_bug.cgi?id=2239213

which could both be solved by just...trying less hard. Let's
not try to set the console layout directly on the old
set_x_keyboard_defaults path, and let's not have special
code to set the console layout when reading configs from the
live image, let's just read in the X layouts from the live
config then rely on the existing code to ask localed to pick
a console layout for us.

This also sets the keyboard model to "pc105" when requesting a
console layout from localed. This is because localed only
applies match "bonuses" for matching variants and options if
the model matches. We need the variant "bonus" match to get
the right result for Bulgarian, at least. Almost every line
in kbd-model-map has the model set to "pc105", so let's just
go with that. The only line with a different model *and* a
variant is es-dvorak, and there is an xkb-converted layout
for that, so it shouldn't make any difference there.

With current stable systemd this will give us wrong results
in several cases (e.g. Russian installs will get US console
layout). With current git HEAD systemd - after
https://github.com/systemd/systemd/pull/29215 - I believe this
gives the same or better results in all cases on both new and
old workflows (I have tested at least Russian and several
Bulgarian paths). With
https://github.com/systemd/systemd/pull/29236 on top, we also
get the correct result for Bulgarian with no variant,
"bg_bds-utf8". Without that PR but with this change, we get the
incorrect "bg_pho-utf8" in that case. Without this change, we
set the console layout to "bg", which does not exist and causes
errors on system startup and fallback to the "us" layout.

Resolves: rhbz#2238854
Resolves: rhbz#2239213

Signed-off-by: Adam Williamson <awilliam@redhat.com>